RepeatingRun CRUD (D.5, D.6)
