+++ app/models/concerns/trip_core.rb
@@ -4,7 +4,7 @@ module TripCore
   extend ActiveSupport::Concern
 
   included do
-    belongs_to :customer, -> { with_deleted }, inverse_of: :trips
+    belongs_to :customer, -> { with_deleted }
     belongs_to :dropoff_address,  -> { with_deleted }, class_name: "Address"
     belongs_to :funding_source, -> { with_deleted }
     belongs_to :mobility, -> { with_deleted }
@@ -32,21 +32,21 @@ module TripCore
 
     accepts_nested_attributes_for :customer
 
-    scope :after,              -> (pickup_time) { where('trips.pickup_time > ?', pickup_time.utc) }
-    scope :after_today,        -> { where('CAST(trips.pickup_time AS date) > ?', Date.today.in_time_zone.utc) }
-    scope :today_and_prior,    -> { where('CAST(trips.pickup_time AS date) <= ?', Date.today.in_time_zone.utc) }
-    scope :prior_to_today,    -> { where('CAST(trips.pickup_time AS date) < ?', Date.today.in_time_zone.utc) }
+    scope :after,              -> (pickup_time) { where('pickup_time > ?', pickup_time.utc) }
+    scope :after_today,        -> { where('CAST(pickup_time AS date) > ?', Date.today.in_time_zone.utc) }
+    scope :today_and_prior,    -> { where('CAST(pickup_time AS date) <= ?', Date.today.in_time_zone.utc) }
+    scope :prior_to_today,    -> { where('CAST(pickup_time AS date) < ?', Date.today.in_time_zone.utc) }
     scope :by_funding_source,  -> (name) { includes(:funding_source).references(:funding_source).where("funding_sources.name = ?", name) }
     scope :by_service_level,   -> (level) { includes(:service_level).references(:service_level).where("service_levels.name = ?", level) }
     scope :by_trip_purpose,    -> (name) { includes(:trip_purpose).references(:trip_purpose).where("trip_purposes.name = ?", name) }
-    scope :during,             -> (pickup_time, appointment_time) { where('NOT ((trips.pickup_time < ? AND trips.appointment_time < ?) OR (trips.pickup_time > ? AND trips.appointment_time > ?))', pickup_time.utc, appointment_time.utc, pickup_time.utc, appointment_time.utc) }
-    scope :for_date,           -> (date) { where('trips.pickup_time >= ? AND trips.pickup_time < ?', date.to_datetime.in_time_zone.utc, date.to_datetime.in_time_zone.utc + 1.day) }
-    scope :for_date_range,     -> (from_date, to_date) { where('trips.pickup_time >= ? AND trips.pickup_time < ?', from_date.to_datetime.in_time_zone.utc, to_date.to_datetime.in_time_zone.utc) }
+    scope :during,             -> (pickup_time, appointment_time) { where('NOT ((pickup_time < ? AND appointment_time < ?) OR (pickup_time > ? AND appointment_time > ?))', pickup_time.utc, appointment_time.utc, pickup_time.utc, appointment_time.utc) }
+    scope :for_date,           -> (date) { where('pickup_time >= ? AND pickup_time < ?', date.to_datetime.in_time_zone.utc, date.to_datetime.in_time_zone.utc + 1.day) }
+    scope :for_date_range,     -> (from_date, to_date) { where('pickup_time >= ? AND pickup_time < ?', from_date.to_datetime.in_time_zone.utc, to_date.to_datetime.in_time_zone.utc) }
     scope :for_provider,       -> (provider_id) { where(provider_id: provider_id) }
     scope :has_scheduled_time, -> { where.not(pickup_time: nil).where.not(appointment_time: nil) }
     scope :individual,         -> { joins(:customer).where(customers: {group: false}) }
     scope :not_called_back,    -> { where('called_back_at IS NULL') }
-    scope :prior_to,           -> (pickup_time) { where('trips.pickup_time < ?', pickup_time.to_datetime.in_time_zone.utc) }
+    scope :prior_to,           -> (pickup_time) { where('pickup_time < ?', pickup_time.to_datetime.in_time_zone.utc) }
   end
 
   def date
