+++ app/views/repeating_runs/_form_header.html.haml
@@ -1,0 +1,19 @@
+.clearfix.col-sm-12#page-header
+  - if edit_mode == :new
+    .pull-right
+      = f.submit translate_helper(:save), class: 'btn action-button'
+      = link_to translate_helper(:cancel), repating_runs_path, class: "btn btn-default"
+    %h1
+      = translate_helper("new_repeating_run")
+  - else
+    .pull-right
+      - if edit_mode == :edit
+        = f.submit translate_helper(:update), class: 'btn action-button'
+        = link_to translate_helper(:cancel), repeating_runs_path, class: "btn btn-default"
+      - else
+        = link_to(translate_helper("delete"), @run, data: {:confirm => translate_helper("delete_run_confirm")}, :method => :delete, :id => 'run_delete', :class=>'btn action-button btn-danger') if !@run.new_record? && can?(:destroy, @run)
+        = link_to translate_helper("edit"), edit_repeating_run_path(@run), :class=>'btn action-button' if !@run.new_record? && can?(:edit, @run)
+
+        = link_to translate_helper(:back), repeating_runs_path, class: "btn btn-default"
+    %h1
+      = translate_helper("repeating_run")
