+++ app/views/repeating_runs/_list.html.haml
@@ -1,0 +1,28 @@
+%table.table.table-striped.table-bordered
+  %thead
+    %tr
+      %th
+      %th= translate_helper("run_name_column")
+      %th= translate_helper("run_date_column")
+      %th= translate_helper("run_start_time_column")
+      %th= translate_helper("run_end_time_column") 
+      %th= translate_helper("run_vehicle_column") 
+      %th= translate_helper("run_driver_column")
+  %tbody
+    - @runs.each do |run|
+      %tr
+        %td
+          = link_to(translate_helper("view"), repeating_run_path(run)) if can?(:show, run)
+        %td
+          = run.name
+        %td
+          = run.date.to_s(:long)
+        %td
+          = run.scheduled_start_time.try(:to_s,:time_only)
+        %td
+          = run.scheduled_end_time.try(:to_s,:time_only)
+        %td
+          = run.vehicle ? run.vehicle.name : translate_helper('no_vehicle')
+        %td
+          = run.driver  ? run.driver.name : translate_helper('no_driver')
+
