+++ app/views/repeating_runs/_repetition_panel.html.haml
@@ -3,5 +3,5 @@
     %h3.panel-title
       = translate_helper(:repetition)
   .panel-body{style: 'padding: 5px;'}
-    = render 'trips/repetition_fields', f: f
+    = render 'repeating_trips/repetition_fields', f: f
