+++ app/views/repeating_runs/edit.html.haml
@@ -1,0 +1,1 @@
+= render 'runs/form', edit_mode: :edit
