+++ app/views/repeating_runs/new.html.haml
@@ -1,0 +1,1 @@
+= render 'runs/form', edit_mode: :edit
