+++ app/views/repeating_runs/show.html.haml
@@ -1,0 +1,4 @@
+= render 'runs/form', edit_mode: :show 
+
+:javascript
+  $(".run_form :input").attr("disabled","disabled");
