+++ app/views/repeating_trips/_form_header.html.haml
@@ -2,7 +2,7 @@
   - if edit_mode == :new
     .pull-right
       = f.submit translate_helper(:save), class: 'btn action-button'
-      = link_to translate_helper(:cancel), trips_path, class: "btn btn-default"
+      = link_to translate_helper(:cancel), repeating_trips_path, class: "btn btn-default"
     %h1
       = translate_helper("new_repeating_trip")
   - else
@@ -14,6 +14,6 @@
         = link_to(translate_helper("delete"), @trip, data: {:confirm => translate_helper("delete_trip_confirm")}, :method => :delete, :id => 'trip_delete', :class=>'btn action-button btn-danger') if !@trip.new_record? && can?(:destroy, @trip)
         = link_to translate_helper("edit"), edit_repeating_trip_path(@trip), :class=>'btn action-button' if !@trip.new_record? && can?(:edit, @trip)
 
-        = link_to translate_helper(:back), trips_path, class: "btn btn-default"
+        = link_to translate_helper(:back), repeating_trips_path, class: "btn btn-default"
     %h1
       = translate_helper("repeating_trip")
