+++ app/views/runs/_date_time_panel.html.haml
@@ -20,16 +20,17 @@
         = f.label :scheduled_end_time, translate_helper("scheduled_end_time")
       .col-sm-9
         = f.text_field :scheduled_end_time, :size => 8, :value => @run.scheduled_end_time.try(:to_s, :time_only), class: 'time-picker form-control', readonly: true
-    .form-group.col-sm-12
-      .col-sm-3
-        = f.label :actual_start_time, translate_helper("actual_start_time")
-      .col-sm-9
-        = f.text_field :actual_start_time, :size => 8, :value => @run.actual_start_time.try(:to_s, :time_only), class: 'time-picker form-control', readonly: true
-    .form-group.col-sm-12
-      .col-sm-3
-        = f.label :actual_end_time, translate_helper("actual_end_time")
-      .col-sm-9
-        = f.text_field :actual_end_time, :size => 8, :value => @run.actual_end_time.try(:to_s, :time_only), class: 'time-picker form-control', readonly: true
+    - unless @is_repeating_run
+      .form-group.col-sm-12
+        .col-sm-3
+          = f.label :actual_start_time, translate_helper("actual_start_time")
+        .col-sm-9
+          = f.text_field :actual_start_time, :size => 8, :value => @run.actual_start_time.try(:to_s, :time_only), class: 'time-picker form-control', readonly: true
+      .form-group.col-sm-12
+        .col-sm-3
+          = f.label :actual_end_time, translate_helper("actual_end_time")
+        .col-sm-9
+          = f.text_field :actual_end_time, :size => 8, :value => @run.actual_end_time.try(:to_s, :time_only), class: 'time-picker form-control', readonly: true
 
 - if edit_mode != :show 
   = render "shared/date_time_picker_javascript"
