--- app/views/runs/_form.html.erb
@@ -1,43 +1,0 @@
-<%= render 'shared/highlight_active_tab_js', is_primary_nav: true, tab_class: 'runs' %>
-<% edit_mode = :edit if !defined? edit_mode%>
-<%= form_for @run, :html => (@remote ? {"data-remote" => true, class: 'run_form'} : {class: 'run_form'} ) do |f| %>
-  <%= render 'form_header', edit_mode: edit_mode, f: f %>
-  <% if @run.errors.any? %>
-    <div id="error_explanation">
-      <h2><%= translate_helper("run_form_error_message", count: @run.errors.count)%></h2>
-
-      <ul>
-      <% @run.errors.full_messages.each do |msg| %>
-        <li><%= msg %></li>
-      <% end %>
-      </ul>
-    </div>
-  <% end %>
-  <div class="col-sm-12">
-    <div class="col-sm-12 col-md-6 form-group">
-      <div class="col-sm-3">
-        <%= f.label :name, translate_helper("name") %>
-      </div>
-      <div class="col-sm-9">
-        <%= f.text_field :name, class: 'form-control' %>
-      </div>
-    </div>
-  </div>
-
-  <div class="col-sm-12 col-md-6">
-    <%= render 'date_time_panel', f: f, edit_mode: edit_mode %>
-  </div>
-  <div class="col-sm-12 col-md-6">
-    <%= render 'repetition_panel', f: f %>
-    <%= render 'vehicle_panel', f: f %>
-    <%= render 'driver_panel', f: f %>
-  </div>
-  <div class="col-sm-12">
-    <% if edit_mode != :new %>
-      <%= render 'trip_list', f: f, edit_mode: edit_mode %>
-    <% end %>
-  </div>
-  
-<% end %>
-
-<%= render 'shared/hide_invisible_form_fields_js', model_name: "run", table_name: 'runs', provider_id:  current_provider.try(:id) %>
