+++ app/views/runs/_form.html.haml
@@ -1,0 +1,42 @@
+:ruby
+  edit_mode = :edit if !defined? edit_mode
+  @is_repeating_run = @run.is_a?(RepeatingRun)
+
+= render 'shared/highlight_active_tab_js', is_primary_nav: true, tab_class: 'runs' 
+
+= form_for @run, :html => (@remote ? {"data-remote" => true, class: 'run_form'} : {class: 'run_form'} ) do |f| 
+  - if @is_repeating_run
+    = render 'repeating_runs/form_header', edit_mode: edit_mode, f: f 
+  - else  
+    = render 'runs/form_header', edit_mode: edit_mode, f: f 
+  - if @run.errors.any? 
+    #error_explanation
+      %h2= translate_helper("run_form_error_message", count: @run.errors.count)
+
+      %ul
+        - @run.errors.full_messages.each do |msg| 
+          %li= msg
+
+  .col-sm-12
+    .col-sm-12.col-md-6.form-group
+      .col-sm-3
+        = f.label :name, translate_helper("name") 
+      .col-sm-9
+        = f.text_field :name, class: 'form-control' 
+
+  .col-sm-12.col-md-6
+    = render 'runs/date_time_panel', f: f, edit_mode: edit_mode
+  .col-sm-12.col-md-6
+    = render 'repeating_runs/repetition_panel', f: f if @is_repeating_run
+    = render 'runs/vehicle_panel', f: f 
+    = render 'runs/driver_panel', f: f 
+
+  - unless @is_repeating_run
+    .col-sm-12
+      - if edit_mode != :new 
+        = render 'runs/trip_list', f: f, edit_mode: edit_mode 
+
+- if @is_repeating_run
+  = render 'shared/hide_invisible_form_fields_js', model_name: "repeatingrun", table_name: 'repeating_runs', provider_id:  current_provider.try(:id) 
+- else
+  = render 'shared/hide_invisible_form_fields_js', model_name: "run", table_name: 'runs', provider_id:  current_provider.try(:id)
