--- app/views/runs/edit.html.erb
@@ -1,1 +1,0 @@
-<%= render 'form', edit_mode: :edit %>
