+++ app/views/runs/edit.html.haml
@@ -1,0 +1,1 @@
+= render 'form', edit_mode: :edit
