--- app/views/runs/new.html.erb
@@ -1,1 +1,0 @@
-<%= render 'form', edit_mode: :new %>
