+++ app/views/runs/new.html.haml
@@ -1,0 +1,1 @@
+= render 'form', edit_mode: :new
