--- app/views/runs/show.html.erb
@@ -1,6 +1,0 @@
-<%= render 'form', edit_mode: :show %>
-
-<script type='text/javascript'>
-  $(".run_form :input").attr("disabled","disabled");
-</script>
-
