+++ app/views/runs/show.html.haml
@@ -1,0 +1,5 @@
+= render 'form', edit_mode: :show 
+
+:javascript
+  $(".run_form :input").attr("disabled","disabled");
+
