+++ config/locales/en.yml
@@ -164,6 +164,11 @@ en:
   repeating_trip: "Subscription Trip"
   repeating_trips_heading: "Subscription Trips"
   new_repeating_trip_text: "Create New Subscription Trip"
+  new_repeating_run: "New Subscription Run"
+  repeating_runs: "View Subscription Runs"
+  repeating_run: "Subscription Run"
+  repeating_runs_heading: "Subscription Runs"
+  new_repeating_run_text: "Create New Subscription Run"
   date: "Date"
   add: "Add"
   edit: "Edit"
